se termino el micro de envios y tramos

+++ tramos/src/main/java/backend/grupo130/tramos/data/repository/PostgresRutaRepositoryI.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 
 
-public interface PostgresRutaRepositoryI extends JpaRepository<RutaTraslado, Integer> {
+public interface PostgresRutaRepositoryI extends JpaRepository<RutaTraslado, Long> {
 
 }
