+++ tramos/src/main/java/backend/grupo130/tramos/dto/ruta/request/RutaGetByIdRequest.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class RutaGetByIdRequest {
 
-    private final Integer idRuta;
+    private final Long idRuta;
 
 }
