+++ tramos/src/main/java/backend/grupo130/tramos/dto/tramo/request/TramoGetByIdRequest.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class TramoGetByIdRequest {
 
-    private final Integer idTramo;
+    private final Long idTramo;
 
 }
