+++ tramos/src/main/java/backend/grupo130/tramos/dto/tramo/request/TramoGetByRutaIdRequest.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class TramoGetByRutaIdRequest {
 
-    private final Integer idRuta;
+    private final Long idRuta;
 
 }
